Split Shooting methods into subpackage
