+++ lib/BoundaryValueDiffEqFIRK/src/BoundaryValueDiffEqFIRK.jl
@@ -22,8 +22,7 @@ import BoundaryValueDiffEq: BoundaryValueDiffEqAlgorithm, BVPJacobianAlgorithm,
                             __maybe_allocate_diffcache, __get_bcresid_prototype, __similar,
                             __vec, __vec_f, __vec_f!, __vec_bc, __vec_bc!,
                             recursive_flatten_twopoint!, __unsafe_nonlinearfunction,
-                            __internal_nlsolve_problem, MaybeDiffCache,
-                            __generate_sparse_jacobian_prototype, __extract_mesh,
+                            __internal_nlsolve_problem, MaybeDiffCache, __extract_mesh,
                             __extract_u0, __has_initial_guess, __initial_guess_length,
                             __initial_guess_on_mesh, __flatten_initial_guess,
                             __build_solution, __Fix3, __sparse_jacobian_cache,
