+++ lib/BoundaryValueDiffEqMIRK/src/BoundaryValueDiffEqMIRK.jl
@@ -22,9 +22,8 @@ import BoundaryValueDiffEq: BoundaryValueDiffEqAlgorithm, BVPJacobianAlgorithm,
                             __maybe_allocate_diffcache, __get_bcresid_prototype, __similar,
                             __vec, __vec_f, __vec_f!, __vec_bc, __vec_bc!,
                             recursive_flatten_twopoint!, __unsafe_nonlinearfunction,
-                            __internal_nlsolve_problem,
-                            __generate_sparse_jacobian_prototype, __extract_mesh,
-                            __extract_u0, __has_initial_guess, __initial_guess_length,
+                            __internal_nlsolve_problem, __extract_mesh, __extract_u0,
+                            __has_initial_guess, __initial_guess_length,
                             __initial_guess_on_mesh, __flatten_initial_guess,
                             __build_solution, __Fix3, __sparse_jacobian_cache,
                             __sparsity_detection_alg, _sparse_like, ColoredMatrix
