+++ lib/BoundaryValueDiffEqShooting/test/nlls_tests.jl
@@ -1,5 +1,5 @@
 @testitem "Overconstrained BVP" begin
-    using LinearAlgebra, JET
+    using BoundaryValueDiffEqShooting, LinearAlgebra, JET
 
     SOLVERS = [
         Shooting(Tsit5(), NewtonRaphson(),
@@ -50,10 +50,10 @@
         @test norm(sol.resid, Inf) < 0.005
 
         JET_SKIP[i] && continue
-        @test_opt target_modules=(BoundaryValueDiffEq,) solve(
+        @test_opt target_modules=(BoundaryValueDiffEqShooting,) solve(
             bvp1, solver; verbose = false, abstol = 1e-6, reltol = 1e-6,
             odesolve_kwargs = (; abstol = 1e-6, reltol = 1e-6)) broken=JET_OPT_BROKEN[i]
-        @test_call target_modules=(BoundaryValueDiffEq,) solve(
+        @test_call target_modules=(BoundaryValueDiffEqShooting,) solve(
             bvp1, solver; verbose = false, abstol = 1e-6, reltol = 1e-6,
             odesolve_kwargs = (; abstol = 1e-6, reltol = 1e-6)) broken=JET_CALL_BROKEN[i]
     end
@@ -87,10 +87,10 @@
         @test norm(sol.resid, Inf) < 0.005
 
         JET_SKIP[i] && continue
-        @test_opt target_modules=(BoundaryValueDiffEq,) solve(
+        @test_opt target_modules=(BoundaryValueDiffEqShooting,) solve(
             bvp2, solver; verbose = false, abstol = 1e-6, reltol = 1e-6,
             odesolve_kwargs = (; abstol = 1e-6, reltol = 1e-6)) broken=JET_OPT_BROKEN[i]
-        @test_call target_modules=(BoundaryValueDiffEq,) solve(
+        @test_call target_modules=(BoundaryValueDiffEqShooting,) solve(
             bvp2, solver; verbose = false, abstol = 1e-6, reltol = 1e-6,
             odesolve_kwargs = (; abstol = 1e-6, reltol = 1e-6)) broken=JET_CALL_BROKEN[i]
     end
@@ -112,10 +112,10 @@
         @test norm(sol.resid, Inf) < 0.009
 
         JET_SKIP[i] && continue
-        @test_opt target_modules=(BoundaryValueDiffEq,) solve(
+        @test_opt target_modules=(BoundaryValueDiffEqShooting,) solve(
             bvp3, solver; verbose = false, abstol = 1e-6, reltol = 1e-6,
             odesolve_kwargs = (; abstol = 1e-6, reltol = 1e-6)) broken=JET_OPT_BROKEN[i]
-        @test_call target_modules=(BoundaryValueDiffEq,) solve(
+        @test_call target_modules=(BoundaryValueDiffEqShooting,) solve(
             bvp3, solver; verbose = false, abstol = 1e-6, reltol = 1e-6,
             odesolve_kwargs = (; abstol = 1e-6, reltol = 1e-6)) broken=JET_CALL_BROKEN[i]
     end
@@ -137,10 +137,10 @@
         @test norm(sol.resid, Inf) < 0.009
 
         JET_SKIP[i] && continue
-        @test_opt target_modules=(BoundaryValueDiffEq,) solve(
+        @test_opt target_modules=(BoundaryValueDiffEqShooting,) solve(
             bvp4, solver; verbose = false, abstol = 1e-6, reltol = 1e-6,
             odesolve_kwargs = (; abstol = 1e-6, reltol = 1e-6)) broken=JET_OPT_BROKEN[i]
-        @test_call target_modules=(BoundaryValueDiffEq,) solve(
+        @test_call target_modules=(BoundaryValueDiffEqShooting,) solve(
             bvp4, solver; verbose = false, abstol = 1e-6, reltol = 1e-6,
             odesolve_kwargs = (; abstol = 1e-6, reltol = 1e-6)) broken=JET_CALL_BROKEN[i]
     end
